requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
requests==2.24.0
gunicorn==20.0.4
psycopg2==2.9.9
typing==3.7.4.3
pydantic==2.6.1
sqlmodel==0.0.16
pydantic_settings==2.2.1
python-jose==3.3.0
cryptography==42.0.4
passlib==1.7.4
bcrypt==4.1.2
email-validator==2.1.0.post1
pytest
pytest-cov
pytest-mock
httpx
typer[all]
python-jose
passlib
python-multipart
alembic
packaging